app/requirements.txt
Requests>=2.21.0
google-cloud-storage
